server/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pandas
numpy
scipy
scikit-learn
requests
streamlit
